requirements.txt
langchain
langchain-core
langchain-groq
langchain-community
langchain-chroma
langgraph
langsmith
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
groq
httpx
requests
python-multipart
aiofiles
pytest
pytest-asyncio
chromadb
crewai
crewai-tools
rank-bm25
pdfplumber
pypdf
sentence-transformers
